user dashbaord,details screens & notification

+++ app/src/main/java/com/techxform/tradintro/feature_main/domain/model/PriceType.kt
@@ -1,7 +1,7 @@
 package com.techxform.tradintro.feature_main.domain.model
 
 data class PriceType(
-    val amount:Double, val type: String
+    val amount:Float, val type: String
 )
 {
